requirements.txt
google-generativeai
google==2.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
langchain==0.2.11
gradio==4.39.0
gradio_client==1.1.1
pdfplumber==0.11.2
python-dotenv